Add files via upload
In this laboratory we will be designing an application, where new activities and handles events are created. It is expected that students can understand the creation of activities and the handling of events emitted by the widgets of an activity.  Follow the instructions in the attached document Laboratory#2.pdf.  (10 points)

Activities creation
Handling events
Manifest Configuration